+++ src/audiodevice.rs
@@ -320,6 +320,7 @@ pub fn new_playback_device(conf: config::Devices) -> Box<dyn PlaybackDevice> {
             target_level: conf.target_level(),
             adjust_period: conf.adjust_period(),
             enable_rate_adjust: conf.rate_adjust(),
+            polling: dev.is_polling(),
         }),
         #[cfg(all(
             feature = "cpal-backend",
@@ -528,6 +529,7 @@ pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             silence_timeout: conf.silence_timeout(),
             stop_on_rate_change: conf.stop_on_rate_change(),
             rate_measure_interval: conf.rate_measure_interval(),
+            polling: dev.is_polling(),
         }),
         #[cfg(all(
             feature = "cpal-backend",